Refactor and reuse the foldOffsetsSizesAndStrides helper function
